pyproject.toml
[project]
name = "bilibili-cli"
version = "0.6.2"
description = "A CLI for Bilibili — browse videos, users, favorites from the terminal"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "jackwener", email = "jakevingoo@gmail.com" }]
keywords = ["bilibili", "cli", "terminal", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "bilibili-api-python>=16.0",
    "click>=8.0",
    "rich>=13.0",
    "aiohttp>=3.0",
    "browser-cookie3>=0.19",
    "pyyaml>=6.0",
    "qrcode>=7.0",
]

[project.optional-dependencies]
audio = [
    "av>=14.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "types-PyYAML>=6.0.12",
    "av>=14.0",
]

[project.urls]
Homepage = "https://github.com/jackwener/bilibili-cli"
Repository = "https://github.com/jackwener/bilibili-cli"
Issues = "https://github.com/jackwener/bilibili-cli/issues"

[project.scripts]
bili = "bili_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bili_cli"]

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false
disable_error_code = ["var-annotated", "arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
markers = [
    "smoke: real-API integration tests (run with: pytest -m smoke)",
]
